resolve comments and port tests
