+++ java/vector/src/main/java/org/apache/arrow/vector/dictionary/DictionaryHashTable.java
@@ -20,6 +20,7 @@ package org.apache.arrow.vector.dictionary;
 import java.util.Objects;
 
 import org.apache.arrow.vector.ValueVector;
+import org.apache.arrow.vector.compare.RangeEqualsVisitor;
 
 /**
  * HashTable used for Dictionary encoding. It holds two vectors (the vector to encode and dictionary vector)
@@ -74,8 +75,6 @@ public class DictionaryHashTable {
 
   private final ValueVector dictionary;
 
-  private final ValueMatcher matcher;
-
   /**
    * Constructs an empty map with the specified initial capacity and load factor.
    */
@@ -96,7 +95,6 @@ public class DictionaryHashTable {
     for (int i = 0; i < this.dictionary.getValueCount(); i++) {
       put(i);
     }
-    this.matcher = new ValueMatcher(dictionary);
   }
 
   public DictionaryHashTable(ValueVector dictionary) {
@@ -143,7 +141,7 @@ public class DictionaryHashTable {
     for (DictionaryHashTable.Entry e = table[index]; e != null ; e = e.next) {
       if ((e.hash == hash)) {
         int dictIndex = e.index;
-        if (matcher.match(dictIndex, toEncode, indexInArray)) {
+        if (toEncode.accept(new RangeEqualsVisitor(dictionary, dictIndex, indexInArray, 1, false))) {
           return dictIndex;
         }
       }
